refactor(config): split config.go into types/defaults/validate/target_override (#151)
