+++ internal/config/config.go
@@ -5,432 +5,12 @@
 package config
 
 import (
-	"errors"
 	"fmt"
-	"net"
-	"net/url"
 	"os"
-	"sort"
 	"strings"
-	"time"
 
 	"gopkg.in/yaml.v3"
 )
-
-// Config is the top-level exporter configuration.
-type Config struct {
-	Listen      ListenConfig      `yaml:"listen"`
-	Discovery   DiscoveryConfig   `yaml:"discovery"`
-	Modules     ModulesConfig     `yaml:"modules"`
-	Credentials CredentialsConfig `yaml:"credentials"`
-	Snapshot    SnapshotConfig    `yaml:"snapshot"`
-	Federation  FederationConfig  `yaml:"federation"`
-	Output      OutputConfig      `yaml:"output"`
-	Targets     []TargetConfig    `yaml:"targets"`
-}
-
-// ListenConfig holds the HTTP/HTTPS listen address and optional auth config.
-// WebConfigFile points at a Prometheus exporter-toolkit web-config YAML — the
-// same schema operators know from snmp_exporter, node_exporter,
-// blackbox_exporter, etc. (see
-// https://github.com/prometheus/exporter-toolkit/blob/master/docs/web-configuration.md).
-// The toolkit handles basic_auth, server TLS, and full mTLS in one file.
-//
-// When both fields are empty the server uses plain HTTP — the default and the
-// canonical Prometheus convention of "scrape from a private network".
-type ListenConfig struct {
-	Addr          string `yaml:"addr"`            // default ":9100"
-	WebConfigFile string `yaml:"web_config_file"` // exporter-toolkit web-config YAML
-
-	// DebugListenAddr, when non-empty, enables a SEPARATE listener serving
-	// net/http/pprof at /debug/pprof/* (issue #69). Empty (the default) means
-	// OFF: no debug listener is created and there is zero runtime overhead. The
-	// debug surface has NO auth and NO TLS — the same model as node_exporter's
-	// debug endpoints — so it MUST NOT be exposed to the internet. Bind it to
-	// localhost or a management interface only (e.g. "127.0.0.1:6060"). It is
-	// never served on the main metrics listener (listen.addr).
-	DebugListenAddr string `yaml:"debug_listen_addr"`
-}
-
-// OutputConfig holds optional output paths beyond Prometheus /metrics.
-type OutputConfig struct {
-	OTLP OTLPOutputConfig `yaml:"otlp"`
-	YANG YANGOutputConfig `yaml:"yang"`
-}
-
-// YANGOutputConfig configures the RFC 8345 YANG-JSON pull endpoint (#75).
-// Disabled by default; when enabled, GET /topology/yang renders the current
-// reconciled topology as RFC 8345/8346 YANG-JSON.
-type YANGOutputConfig struct {
-	Enabled   bool   `yaml:"enabled"`    // default false
-	NetworkID string `yaml:"network_id"` // RFC 8345 network-id; default applied in defaults
-}
-
-// OTLPOutputConfig configures the OTLP push output.
-type OTLPOutputConfig struct {
-	Enabled         bool          `yaml:"enabled"`
-	Endpoint        string        `yaml:"endpoint"`
-	Timeout         time.Duration `yaml:"timeout"`
-	HeartbeatCycles int           `yaml:"heartbeat_cycles"`
-
-	// Protocol selects the OTLP transport: "http" (default) or "grpc".
-	// Empty preserves the pre-SDK behaviour (OTLP/HTTP). The payload encoding
-	// is always protobuf — the OpenTelemetry Go SDK exporters do not implement
-	// OTLP/JSON. This is a wire-format change from the pre-v1.5.0 hand-rolled
-	// JSON path; OTLP receivers must accept protobuf (the OTLP default).
-	Protocol OTLPProtocol `yaml:"protocol"`
-
-	// Traces configures opt-in OpenTelemetry tracing of the exporter's own
-	// discovery cycle (issue #68). It reuses Endpoint, Protocol, and auth from
-	// this block — tracing has no endpoint of its own.
-	Traces OTLPTracesConfig `yaml:"traces"`
-}
-
-// OTLPTracesConfig configures the opt-in trace signal layered on the OTLP
-// output. Tracing is disabled by default; when enabled it exports spans of the
-// discovery cycle over the same Endpoint/Protocol as OTLP metrics and logs.
-type OTLPTracesConfig struct {
-	// Enabled turns on tracing of the discovery cycle. Default false: when
-	// off, instrumentation calls resolve to the OTel no-op tracer and emit
-	// nothing.
-	Enabled bool `yaml:"enabled"`
-	// SampleRate is the head-sampling ratio in [0,1] applied to the root
-	// discovery.cycle span; child spans inherit the parent decision
-	// (ParentBased). Default 0.1. Set to 1.0 to sample every cycle (useful for
-	// low-cardinality debugging) or to a small fraction in production.
-	//
-	// A pointer so a config that omits sample_rate gets the 0.1 default while a
-	// config that explicitly sets sample_rate: 0.0 (sample nothing) is honoured
-	// rather than silently overridden by the default.
-	SampleRate *float64 `yaml:"sample_rate"`
-}
-
-// FederationConfig is the LD-15–LD-20 multi-instance coordination config.
-// Role selects the operating mode; the remaining fields are mode-specific.
-type FederationConfig struct {
-	// Role selects the operating mode. Default "standalone" preserves existing
-	// single-instance behaviour. "uncoordinated" adds boundary-observation
-	// metrics for Mimir recording-rule stitching (LD-15). "spoke" and "hub"
-	// configure the H-PCE-style push hierarchy (LD-16).
-	Role Role `yaml:"role"` // standalone | uncoordinated | spoke | hub
-
-	// SpokeTimeout governs spoke eviction on the hub per LD-18: a spoke that
-	// has not pushed within this duration has its edges removed from the
-	// aggregated graph. Defaults to 3× discovery.interval.
-	SpokeTimeout time.Duration `yaml:"spoke_timeout"`
-
-	// KnownInterDomainLinks is the LD-19 static stitching override. Each
-	// entry is a fully-qualified boundary link tuple; the hub injects these as
-	// confirmed bidirectional edges regardless of automatic name-matching.
-	KnownInterDomainLinks []InterDomainLink `yaml:"known_inter_domain_links"`
-
-	// Hub holds the hub-side federation server settings (role: hub).
-	Hub FederationHubConfig `yaml:"hub"`
-
-	// Spoke holds the spoke-side settings (role: spoke).
-	Spoke FederationSpokeConfig `yaml:"spoke"`
-}
-
-// FederationHubConfig holds the hub's federation server settings.
-// TLS fields are file paths, not inline PEM. Per LD-20, all three must be set.
-type FederationHubConfig struct {
-	// ListenAddr is the address the hub listens on for spoke pushes.
-	// Separate from web.listen-address so Prometheus scrapes do not require mTLS.
-	ListenAddr string `yaml:"listen_addr"`
-	TLSCACert  string `yaml:"tls_ca_cert"` // path to PEM CA certificate for spoke client verification
-	TLSCert    string `yaml:"tls_cert"`    // path to PEM server certificate
-	TLSKey     string `yaml:"tls_key"`     // path to PEM server private key
-	// MaxGraphEdges, if > 0, rejects combined-graph updates with more edges than
-	// this limit. Protects scrape latency and memory from runaway topologies.
-	MaxGraphEdges int `yaml:"max_graph_edges"`
-	// MaxGraphDevices, if > 0, rejects combined-graph updates with more devices
-	// than this limit. Protects scrape latency and memory from runaway topologies.
-	MaxGraphDevices int `yaml:"max_graph_devices"`
-	// LooseDeviceNameMatching enables domain-suffix stripping during OOS
-	// neighbour matching. The zero value (false) is the safe default since
-	// v1.3.0: "core-sw.dc1.example.com" and "core-sw.dc2.example.com" remain
-	// distinct. Set to true to restore the pre-v1.3.0 behaviour for single-site
-	// deployments where short and FQDN forms of the same device must reconcile
-	// to one node — both forms normalise to "core-sw".
-	LooseDeviceNameMatching bool `yaml:"loose_device_name_matching"`
-	// MinPushInterval rejects per-spoke pushes received sooner than this
-	// duration after the previous accepted push from the same spoke_id, with
-	// a 429 Too Many Requests response and a Retry-After header. Defense in
-	// depth on top of mTLS to bound the cost of a misconfigured or malicious
-	// spoke that ignores its own push backoff. 0 (default) disables the
-	// check; set to roughly half the spoke's discovery_interval for a sane
-	// floor. Must be strictly less than spoke_timeout.
-	MinPushInterval time.Duration `yaml:"min_push_interval"`
-}
-
-// FederationSpokeConfig holds the spoke's settings.
-// TLS fields are file paths, not inline PEM. Per LD-20, all three must be set.
-type FederationSpokeConfig struct {
-	// SpokeID is a unique identifier for this spoke instance. Used as the key
-	// in the hub's spoke-edge store and in federation metric labels.
-	SpokeID string `yaml:"spoke_id"`
-	// HubURL is the base URL of the hub federation server, e.g. https://hub:9101.
-	HubURL    string `yaml:"hub_url"`
-	TLSCACert string `yaml:"tls_ca_cert"` // path to PEM CA certificate for hub server verification
-	TLSCert   string `yaml:"tls_cert"`    // path to PEM spoke client certificate
-	TLSKey    string `yaml:"tls_key"`     // path to PEM spoke client private key
-}
-
-// InterDomainLink is one LD-19 static boundary-link override. LocalDevice,
-// LocalPort, RemoteDevice, and RemotePort are required. LinkKind defaults to
-// "ethernet" when omitted.
-type InterDomainLink struct {
-	LocalDevice  string `yaml:"local_device"`
-	LocalPort    string `yaml:"local_port"`
-	RemoteDevice string `yaml:"remote_device"`
-	RemotePort   string `yaml:"remote_port"`
-	LinkKind     string `yaml:"link_kind,omitempty"` // defaults to "ethernet"
-}
-
-// DiscoveryConfig controls the global discovery cycle.
-//
-// Scope (LD-11): CIDRAllowList is a hard bound on what the exporter polls.
-// Targets must fall within it; LLDP/CDP-discovered neighbours outside it
-// surface as network_topology_out_of_scope_neighbour and are never polled.
-//
-// UnconfirmedLinkTTLCycles (LD-14): a unidirectional link removed from the
-// graph after this many consecutive unconfirmed cycles. Default 3.
-type DiscoveryConfig struct {
-	Interval                 time.Duration `yaml:"interval"`
-	TimeoutPerDevice         time.Duration `yaml:"timeout_per_device"`
-	Parallelism              int           `yaml:"parallelism"`
-	Scope                    ScopeConfig   `yaml:"scope"`
-	UnconfirmedLinkTTLCycles int           `yaml:"unconfirmed_link_ttl_cycles"`
-
-	// CycleBudgetFraction caps the discovery cycle at this fraction of Interval.
-	// Goroutines still running when the budget expires are cancelled via context.
-	// Default 0.8 (80% of Interval). Must be in (0, 1].
-	CycleBudgetFraction float64 `yaml:"cycle_budget_fraction"`
-
-	// TimeoutPerModule caps each individual module walk within a device's
-	// per-device timeout. 0 means no additional cap (default). When set, a
-	// slow LLDP walk cannot consume the full timeout_per_device for all modules.
-	TimeoutPerModule time.Duration `yaml:"timeout_per_module"`
-
-	// MaxGraphDevices, if > 0, rejects local graph updates with more devices
-	// than this value. Mirrors FederationHubConfig.MaxGraphDevices for
-	// standalone and spoke mode.
-	MaxGraphDevices int `yaml:"max_graph_devices"`
-
-	// MaxGraphEdges, if > 0, rejects local graph updates with more edges than
-	// this value. Mirrors FederationHubConfig.MaxGraphEdges for standalone
-	// and spoke mode.
-	MaxGraphEdges int `yaml:"max_graph_edges"`
-
-	// PerTargetPDURatePerSecond caps the steady-state SNMP request (PDU) rate
-	// issued against a single device during a discovery cycle. A misconfigured
-	// high parallelism with all modules enabled can otherwise drive hundreds of
-	// PDUs/sec at one target and self-DoS its SNMP daemon (issue #72). The cap
-	// is applied per device per cycle (per-target isolation — limiters are not
-	// shared across devices). 0 (default) means unlimited, preserving today's
-	// behaviour. Must be >= 0.
-	PerTargetPDURatePerSecond int `yaml:"per_target_pdu_rate_per_second"`
-
-	// TargetOverrides scopes discovery to a subset of the globally-enabled
-	// modules on a per-CIDR basis (issue #74). For an IP matching an override's
-	// cidr, ONLY the listed modules run (intersected with modules.<m>.enabled);
-	// an IP matching no override runs ALL enabled modules — today's behaviour,
-	// unchanged. Heterogeneous fleets use this so e.g. BGP is not walked against
-	// an access switch that does not speak BGP (avoiding a guaranteed
-	// mib_unimplemented walker outcome). Most-specific CIDR wins (longest prefix
-	// match), the same precedence model as credentials.assignments.
-	TargetOverrides []TargetOverride `yaml:"target_overrides"`
-
-	// overrideResolver is the parsed, longest-prefix-sorted form of
-	// TargetOverrides, built once during validate() so the discovery loop never
-	// re-parses CIDRs per device per cycle. nil when no overrides are
-	// configured (the unchanged default path).
-	overrideResolver *targetOverrideResolver
-
-	// SNMP holds discovery-layer SNMP transport tuning that is not per-module
-	// (the per-module on/off toggles live under modules.snmp). Currently this is
-	// just the opt-in session pool (issue #83).
-	SNMP DiscoverySNMPConfig `yaml:"snmp"`
-
-	// LivenessMaxStaleCycles gates the /healthz liveness probe on discovery
-	// staleness. When the most recent cycle is older than
-	// interval × liveness_max_stale_cycles, /healthz returns 503 so Kubernetes
-	// restarts a process whose discovery loop has wedged (a deadlocked loop can
-	// no longer update its own last-cycle timestamp). Default 3. A value of 0
-	// disables the gate, restoring the prior behaviour where /healthz is always
-	// 200 once the process is up. The gate is inert in pure-hub mode regardless
-	// of this value, since a hub runs no local discovery loop.
-	//
-	// A pointer so the loader can distinguish "unset" (nil → default 3) from an
-	// explicit 0 (disable the gate). After applyDefaults this is always non-nil;
-	// read it via LivenessMaxStaleCyclesValue. Must be >= 0.
-	LivenessMaxStaleCycles *int `yaml:"liveness_max_stale_cycles"`
-}
-
-// LivenessMaxStaleCyclesValue returns the resolved liveness-gate cycle count.
-// applyDefaults guarantees the field is non-nil, but this accessor is nil-safe
-// (returns 0 → gate disabled) so callers never dereference a nil pointer.
-func (d DiscoveryConfig) LivenessMaxStaleCyclesValue() int {
-	if d.LivenessMaxStaleCycles == nil {
-		return 0
-	}
-	return *d.LivenessMaxStaleCycles
-}
-
-// DiscoverySNMPConfig groups discovery-wide SNMP transport options.
-type DiscoverySNMPConfig struct {
-	SessionPool SessionPoolConfig `yaml:"session_pool"`
-}
-
-// SessionPoolConfig configures the opt-in per-target SNMP session pool (issue
-// #83). When disabled (the default), every (target × module) walk opens and
-// closes a fresh UDP session, exactly as before. When enabled, a target reuses
-// one session across its modules within a cycle and across cycles, keyed by
-// (target IP, credential profile), cutting socket / conntrack churn on large
-// fleets (see docs/operator/scale.md). The session is bounded at one per
-// (target × profile) (~50KB each), so memory cost scales with fleet size, not
-// module count.
-type SessionPoolConfig struct {
-	// Enabled turns the pool on. Default false: behaviour is byte-identical to
-	// pre-#83 (fresh session per walk).
-	Enabled bool `yaml:"enabled"`
-
-	// MaxIdle is how long a pooled session may sit unused before the background
-	// evictor closes it. 0 (the default) means "5 × discovery.interval",
-	// computed in applyDefaults. Must be >= 0. Closing an idle session also
-	// clears the credentials it held (issue #5 retention bound).
-	MaxIdle time.Duration `yaml:"max_idle"`
-}
-
-// TargetOverride restricts the modules that run against IPs inside CIDR to the
-// listed Modules (issue #74). See DiscoveryConfig.TargetOverrides for the full
-// semantics. Each module name must be one of the canonical module identifiers
-// (lldp, cdp, fdb, ospf, bgp, isis, mpls_te) and must be enabled globally; an
-// empty Modules list is rejected as ambiguous.
-type TargetOverride struct {
-	CIDR    string   `yaml:"cidr"`
-	Modules []string `yaml:"modules"`
-}
-
-// ScopeConfig is the LD-11 polling-scope guard. CIDRAllowList is mandatory
-// when any module is enabled; the empty list is treated as a misconfiguration
-// rather than as "poll everything".
-type ScopeConfig struct {
-	CIDRAllowList []string `yaml:"cidr_allow_list"`
-}
-
-// FDBConfig holds FDB-module-specific tuning knobs. MaxVlans caps the number
-// of VLANs that walkVlanCommunityFdbs will iterate on IOS devices; beyond
-// that limit the walk is truncated and a warning is logged.
-type FDBConfig struct {
-	Enabled  bool `yaml:"enabled"`
-	MaxVlans int  `yaml:"max_vlans"`
-}
-
-// BGPConfig holds BGP-module-specific tuning. DisableV2MIB is the kill-switch
-// for the BGP4-V2-MIB / vendor-specific peer-table walkers that surface IPv6
-// BGP adjacencies; RFC 4273 BGP4-MIB is IPv4-only. The zero value (false) is
-// the safe default since v1.3.0: v2 walkers run. Set to true to fall back to
-// RFC 4273-only behaviour if a vendor regression appears.
-type BGPConfig struct {
-	Enabled      bool `yaml:"enabled"`
-	DisableV2MIB bool `yaml:"disable_v2_mib"`
-}
-
-// ModulesConfig toggles individual discovery modules. Each module's spec
-// citation lives in the corresponding internal/discovery/<name>/<name>.go header.
-type ModulesConfig struct {
-	SNMP ModuleSNMP   `yaml:"snmp"`
-	LLDP ModuleToggle `yaml:"lldp"`
-	CDP  ModuleToggle `yaml:"cdp"`
-	BGP  BGPConfig    `yaml:"bgp"`
-	OSPF ModuleToggle `yaml:"ospf"`
-	// ARP is enrichment, not an edge source — it walks ipNetToMediaTable to
-	// build a MAC→IP map used to backfill DstPorts on FDB-derived edges
-	// when LLDP/CDP do not name the neighbour. Defaults to disabled, consistent
-	// with every other module toggle; only useful when modules.fdb.enabled is
-	// also true. main emits a startup warning when FDB is on and ARP is off.
-	ARP    ModuleToggle `yaml:"arp"`
-	FDB    FDBConfig    `yaml:"fdb"`
-	ISIS   ModuleToggle `yaml:"isis"`
-	MPLSTE ModuleToggle `yaml:"mpls_te"`
-}
-
-// ModuleToggle is the minimum module config: enabled / disabled.
-type ModuleToggle struct {
-	Enabled bool `yaml:"enabled"`
-}
-
-// ModuleSNMP toggles the SNMP transport. Per-device credential resolution
-// lives under credentials: (LD-12). The legacy CommunityEnv field is retained
-// for the dev-time single-profile path; production deployments should use
-// credentials.profiles.
-type ModuleSNMP struct {
-	Enabled      bool        `yaml:"enabled"`
-	Version      SNMPVersion `yaml:"version"` // v2c | v3
-	CommunityEnv string      `yaml:"community_env"`
-}
-
-// CredentialsConfig encodes LD-12: named profiles, per-device assignments,
-// ordered fallback, and a global trial rate limit.
-//
-// Secret values are dereferenced from environment variables only — every
-// *Env field carries the variable name, never the secret itself.
-type CredentialsConfig struct {
-	Profiles           []CredentialProfile    `yaml:"profiles"`
-	Assignments        []CredentialAssignment `yaml:"assignments"`
-	FallbackOrder      []string               `yaml:"fallback_order"`
-	TrialRatePerSecond int                    `yaml:"trial_rate_per_second"`
-}
-
-// Profile type constants for CredentialProfile.Type are defined as the named
-// ProfileType enum in enums.go.
-
-// CredentialProfile is one named credential the exporter can try against
-// a device. Type selects which *Env fields are consulted.
-type CredentialProfile struct {
-	Name         string      `yaml:"name"`
-	Type         ProfileType `yaml:"type"` // snmp_v2c | snmp_v3
-	CommunityEnv string      `yaml:"community_env,omitempty"`
-	UsernameEnv  string      `yaml:"username_env,omitempty"`
-	AuthProtocol string      `yaml:"auth_protocol,omitempty"` // SHA (recommended) | SHA-256 | SHA-384 | SHA-512 | MD5 (deprecated, broken)
-	AuthKeyEnv   string      `yaml:"auth_key_env,omitempty"`
-	PrivProtocol string      `yaml:"priv_protocol,omitempty"` // AES (recommended) | AES-192 | AES-256 | DES (deprecated, broken)
-	PrivKeyEnv   string      `yaml:"priv_key_env,omitempty"`
-	// Retries is the number of SNMP retries per request. Defaults to 1.
-	Retries int `yaml:"retries"`
-	// ContextName is the SNMPv3 context name. Empty string means no context (default).
-	ContextName string `yaml:"context_name,omitempty"`
-}
-
-// CredentialAssignment binds one or more profiles to a device or CIDR.
-// Resolution order in the resolver: explicit IP > most-specific CIDR >
-// fallback_order.
-type CredentialAssignment struct {
-	IP       string   `yaml:"ip,omitempty"`
-	CIDR     string   `yaml:"cidr,omitempty"`
-	Profiles []string `yaml:"profiles"`
-}
-
-// SnapshotConfig is the LD-13 graph-persistence path. The exporter loads
-// the snapshot on startup and serves stale-but-valid metrics until the first
-// live cycle completes. Snapshot is always active; set Path to override the
-// default location.
-type SnapshotConfig struct {
-	Path string `yaml:"path"`
-}
-
-// TargetConfig provides per-device enrichment: SNMP port override, site label,
-// and free-form labels applied to the discovered device record. The CIDR
-// allow-list in discovery.scope drives what gets polled; targets are matched
-// against discovered IPs to attach this metadata. A target whose IP falls
-// outside the allow-list is rejected at startup (LD-11).
-type TargetConfig struct {
-	Host   string            `yaml:"host"`
-	Port   int               `yaml:"port"`
-	Site   string            `yaml:"site"`
-	Labels map[string]string `yaml:"labels"`
-}
 
 // Load parses the configuration file at path and applies defaults.
 // Unknown YAML keys are rejected (KnownFields mode) so that removed
@@ -452,654 +32,4 @@ func Load(path string) (*Config, error) {
 		return nil, err
 	}
 	return &c, nil
-}
-
-func (c *Config) applyDefaults() {
-	if c.Listen.Addr == "" {
-		c.Listen.Addr = ":9100"
-	}
-	if c.Discovery.Interval == 0 {
-		c.Discovery.Interval = 60 * time.Second
-	}
-	if c.Discovery.TimeoutPerDevice == 0 {
-		c.Discovery.TimeoutPerDevice = 10 * time.Second
-	}
-	if c.Discovery.Parallelism == 0 {
-		c.Discovery.Parallelism = 32
-	}
-	if c.Discovery.UnconfirmedLinkTTLCycles == 0 {
-		c.Discovery.UnconfirmedLinkTTLCycles = 3
-	}
-	if c.Discovery.CycleBudgetFraction == 0 {
-		c.Discovery.CycleBudgetFraction = 0.8
-	}
-	// Liveness staleness gate defaults to 3 cycles. The field is a pointer so an
-	// unset value (nil) gets the default 3, while an explicit 0 in YAML is
-	// preserved as "disable the gate" — the two are otherwise indistinguishable
-	// for a plain int. See DiscoveryConfig.LivenessMaxStaleCycles.
-	if c.Discovery.LivenessMaxStaleCycles == nil {
-		def := 3
-		c.Discovery.LivenessMaxStaleCycles = &def
-	}
-	// Session-pool idle TTL defaults to 5 × interval (issue #83). Computed after
-	// the interval default above so a fully-defaulted config still gets a sane
-	// non-zero TTL. Only meaningful when the pool is enabled, but harmless to set
-	// regardless.
-	if c.Discovery.SNMP.SessionPool.MaxIdle == 0 {
-		c.Discovery.SNMP.SessionPool.MaxIdle = 5 * c.Discovery.Interval
-	}
-	if c.Modules.SNMP.Version == "" {
-		c.Modules.SNMP.Version = SNMPVersionV2c
-	}
-	if c.Credentials.TrialRatePerSecond == 0 {
-		c.Credentials.TrialRatePerSecond = 5
-	}
-	for i := range c.Credentials.Profiles {
-		if c.Credentials.Profiles[i].Retries == 0 {
-			c.Credentials.Profiles[i].Retries = 1
-		}
-	}
-	if c.Modules.FDB.MaxVlans == 0 {
-		c.Modules.FDB.MaxVlans = 100
-	}
-	if c.Snapshot.Path == "" {
-		c.Snapshot.Path = "/var/lib/network-topology-exporter/snapshot.json"
-	}
-	for i := range c.Targets {
-		if c.Targets[i].Port == 0 {
-			c.Targets[i].Port = 161
-		}
-	}
-	if c.Federation.Role == "" {
-		c.Federation.Role = RoleStandalone
-	}
-	if c.Federation.SpokeTimeout == 0 {
-		c.Federation.SpokeTimeout = 3 * c.Discovery.Interval
-	}
-	if c.Federation.Hub.ListenAddr == "" {
-		c.Federation.Hub.ListenAddr = ":9101"
-	}
-	if c.Output.OTLP.HeartbeatCycles == 0 {
-		c.Output.OTLP.HeartbeatCycles = 10
-	}
-	if c.Output.OTLP.Timeout == 0 {
-		c.Output.OTLP.Timeout = 10 * time.Second
-	}
-	if c.Output.OTLP.Protocol == "" {
-		c.Output.OTLP.Protocol = OTLPProtocolHTTP
-	}
-	if c.Output.OTLP.Traces.SampleRate == nil {
-		def := 0.1
-		c.Output.OTLP.Traces.SampleRate = &def
-	}
-	if c.Output.YANG.NetworkID == "" {
-		c.Output.YANG.NetworkID = "network-topology-exporter"
-	}
-}
-
-func (c *Config) validateListen() error {
-	if c.Listen.WebConfigFile != "" {
-		if _, err := os.Stat(c.Listen.WebConfigFile); err != nil {
-			return fmt.Errorf("listen.web_config_file %q: %w", c.Listen.WebConfigFile, err)
-		}
-	}
-	return nil
-}
-
-func (c *Config) validate() error {
-	if err := c.validateListen(); err != nil {
-		return err
-	}
-	if c.Discovery.Parallelism < 1 {
-		return errors.New("discovery.parallelism must be >= 1")
-	}
-	if c.Discovery.UnconfirmedLinkTTLCycles < 1 {
-		return errors.New("discovery.unconfirmed_link_ttl_cycles must be >= 1")
-	}
-	if c.Discovery.CycleBudgetFraction <= 0 || c.Discovery.CycleBudgetFraction > 1 {
-		return errors.New("discovery.cycle_budget_fraction must be in (0, 1]")
-	}
-	if c.Discovery.TimeoutPerModule < 0 {
-		return errors.New("discovery.timeout_per_module must be >= 0")
-	}
-	if c.Discovery.TimeoutPerModule > 0 && c.Discovery.TimeoutPerModule >= c.Discovery.TimeoutPerDevice {
-		return errors.New("discovery.timeout_per_module must be less than discovery.timeout_per_device")
-	}
-	if c.Discovery.MaxGraphDevices < 0 {
-		return errors.New("discovery.max_graph_devices must be >= 0 (0 = unlimited)")
-	}
-	if c.Discovery.MaxGraphEdges < 0 {
-		return errors.New("discovery.max_graph_edges must be >= 0 (0 = unlimited)")
-	}
-	if c.Discovery.PerTargetPDURatePerSecond < 0 {
-		return errors.New("discovery.per_target_pdu_rate_per_second must be >= 0 (0 = unlimited)")
-	}
-	if c.Discovery.SNMP.SessionPool.MaxIdle < 0 {
-		return errors.New("discovery.snmp.session_pool.max_idle must be >= 0 (0 = default 5× interval)")
-	}
-	if c.Discovery.Interval <= 0 {
-		return errors.New("discovery.interval must be > 0")
-	}
-	if c.Discovery.LivenessMaxStaleCyclesValue() < 0 {
-		return errors.New("discovery.liveness_max_stale_cycles must be >= 0 (0 = liveness staleness gate disabled)")
-	}
-	if c.Discovery.TimeoutPerDevice <= 0 {
-		return errors.New("discovery.timeout_per_device must be > 0")
-	}
-	if err := c.validateScope(); err != nil {
-		return err
-	}
-	if err := c.validateTargetOverrides(); err != nil {
-		return err
-	}
-	if c.Modules.SNMP.Enabled {
-		switch c.Modules.SNMP.Version {
-		case SNMPVersionV2c, SNMPVersionV3:
-		default:
-			return fmt.Errorf("modules.snmp.version must be v2c or v3, got %q", c.Modules.SNMP.Version)
-		}
-	}
-	if err := c.validateCredentials(); err != nil {
-		return err
-	}
-	if err := c.validateTargets(); err != nil {
-		return err
-	}
-	if err := c.validateFDB(); err != nil {
-		return err
-	}
-	if err := c.validateFederation(); err != nil {
-		return err
-	}
-	if c.Output.OTLP.Timeout < 0 {
-		return fmt.Errorf("output.otlp.timeout must be >= 0 (0 uses the default)")
-	}
-	if c.Output.OTLP.HeartbeatCycles < 1 {
-		return errors.New("output.otlp.heartbeat_cycles must be >= 1")
-	}
-	switch c.Output.OTLP.Protocol {
-	case "", OTLPProtocolHTTP, OTLPProtocolGRPC:
-	default:
-		return fmt.Errorf("output.otlp.protocol must be http or grpc, got %q", c.Output.OTLP.Protocol)
-	}
-	if sr := c.Output.OTLP.Traces.SampleRate; sr != nil && (*sr < 0 || *sr > 1) {
-		return fmt.Errorf("output.otlp.traces.sample_rate must be in [0, 1], got %v", *sr)
-	}
-	if c.Output.OTLP.Enabled {
-		// gRPC endpoints are bare host:port authorities (or a URL); HTTP
-		// endpoints must be a full http/https URL as before.
-		if c.Output.OTLP.Protocol == OTLPProtocolGRPC {
-			if c.Output.OTLP.Endpoint == "" {
-				return errors.New("output.otlp.endpoint: endpoint is required")
-			}
-		} else if err := validateHTTPEndpoint(c.Output.OTLP.Endpoint); err != nil {
-			return fmt.Errorf("output.otlp.endpoint: %w", err)
-		}
-	}
-	return nil
-}
-
-func validateHTTPEndpoint(endpoint string) error {
-	if endpoint == "" {
-		return errors.New("endpoint is required")
-	}
-	u, err := url.Parse(endpoint)
-	if err != nil {
-		return fmt.Errorf("invalid URL: %w", err)
-	}
-	if u.Scheme != "http" && u.Scheme != "https" {
-		return fmt.Errorf("scheme must be http or https, got %q", u.Scheme)
-	}
-	if u.Host == "" {
-		return errors.New("host is required")
-	}
-	return nil
-}
-
-func validateHTTPSEndpoint(endpoint string) error {
-	if endpoint == "" {
-		return errors.New("endpoint is required")
-	}
-	u, err := url.Parse(endpoint)
-	if err != nil {
-		return fmt.Errorf("invalid URL: %w", err)
-	}
-	if u.Scheme != "https" {
-		return fmt.Errorf("scheme must be https, got %q", u.Scheme)
-	}
-	if u.Host == "" {
-		return errors.New("host is required")
-	}
-	return nil
-}
-
-// validateScope enforces LD-11: the allow-list is mandatory, every entry
-// must parse as a CIDR, and every target host must resolve into one of the
-// allow-list networks.
-func (c *Config) validateScope() error {
-	if len(c.Targets) == 0 {
-		return nil // dev-time: no targets, no scope check needed.
-	}
-	if len(c.Discovery.Scope.CIDRAllowList) == 0 {
-		return errors.New("discovery.scope.cidr_allow_list is required when targets are defined (LD-11)")
-	}
-	nets := make([]*net.IPNet, 0, len(c.Discovery.Scope.CIDRAllowList))
-	for _, raw := range c.Discovery.Scope.CIDRAllowList {
-		_, n, err := net.ParseCIDR(raw)
-		if err != nil {
-			return fmt.Errorf("discovery.scope.cidr_allow_list entry %q: %w", raw, err)
-		}
-		nets = append(nets, n)
-	}
-	for _, t := range c.Targets {
-		ip := net.ParseIP(t.Host)
-		if ip == nil {
-			// Hostname rather than IP; defer to runtime resolution. Scope is
-			// enforced again at poll time once the hostname resolves.
-			continue
-		}
-		if !cidrContainsAny(nets, ip) {
-			return fmt.Errorf("target %q is outside discovery.scope.cidr_allow_list (LD-11)", t.Host)
-		}
-	}
-	return nil
-}
-
-// validateCredentials checks LD-12 invariants: profile names are unique,
-// every assignment / fallback name resolves to a defined profile, and every
-// profile carries the env-var fields its type requires.
-func (c *Config) validateCredentials() error {
-	if c.Credentials.TrialRatePerSecond < 1 {
-		return errors.New("credentials.trial_rate_per_second must be >= 1")
-	}
-	if len(c.Credentials.Profiles) == 0 {
-		return nil // legacy single-community path; ModuleSNMP.CommunityEnv applies.
-	}
-	known := make(map[string]CredentialProfile, len(c.Credentials.Profiles))
-	for i := range c.Credentials.Profiles {
-		p := c.Credentials.Profiles[i]
-		if p.Name == "" {
-			return errors.New("credentials.profiles[].name is required")
-		}
-		if _, dup := known[p.Name]; dup {
-			return fmt.Errorf("credentials.profiles: duplicate name %q", p.Name)
-		}
-		if p.Retries < 0 {
-			return fmt.Errorf("profile %q: retries must be >= 0", p.Name)
-		}
-		switch p.Type {
-		case ProfileTypeSNMPv2c:
-			if p.CommunityEnv == "" {
-				return fmt.Errorf("profile %q (snmp_v2c) requires community_env", p.Name)
-			}
-		case ProfileTypeSNMPv3:
-			if p.UsernameEnv == "" {
-				return fmt.Errorf("profile %q (snmp_v3) requires username_env", p.Name)
-			}
-			authProto, err := normalizeAuthProtocol(p.AuthProtocol)
-			if err != nil {
-				return fmt.Errorf("profile %q: %w", p.Name, err)
-			}
-			privProto, err := normalizePrivProtocol(p.PrivProtocol)
-			if err != nil {
-				return fmt.Errorf("profile %q: %w", p.Name, err)
-			}
-			if authProto != "" && p.AuthKeyEnv == "" {
-				return fmt.Errorf("snmpv3 profile %q: auth_protocol requires auth_key_env", p.Name)
-			}
-			if privProto != "" && p.PrivKeyEnv == "" {
-				return fmt.Errorf("snmpv3 profile %q: priv_protocol requires priv_key_env", p.Name)
-			}
-			p.AuthProtocol = authProto
-			p.PrivProtocol = privProto
-		default:
-			return fmt.Errorf("profile %q has unknown type %q", p.Name, p.Type)
-		}
-		c.Credentials.Profiles[i] = p
-		known[p.Name] = p
-	}
-	for i, a := range c.Credentials.Assignments {
-		if a.IP == "" && a.CIDR == "" {
-			return errors.New("credentials.assignments[]: ip or cidr is required")
-		}
-		if a.IP != "" && a.CIDR != "" {
-			return fmt.Errorf("credentials.assignments[%d]: ip and cidr are mutually exclusive", i)
-		}
-		if a.IP != "" && net.ParseIP(a.IP) == nil {
-			return fmt.Errorf("credentials.assignments: invalid ip %q", a.IP)
-		}
-		if a.CIDR != "" {
-			if _, _, err := net.ParseCIDR(a.CIDR); err != nil {
-				return fmt.Errorf("credentials.assignments: invalid cidr %q: %w", a.CIDR, err)
-			}
-		}
-		for _, n := range a.Profiles {
-			if _, ok := known[n]; !ok {
-				return fmt.Errorf("credentials.assignments references unknown profile %q", n)
-			}
-		}
-	}
-	for _, n := range c.Credentials.FallbackOrder {
-		if _, ok := known[n]; !ok {
-			return fmt.Errorf("credentials.fallback_order references unknown profile %q", n)
-		}
-	}
-	return nil
-}
-
-func normalizeAuthProtocol(raw string) (string, error) {
-	switch v := strings.ToUpper(strings.TrimSpace(raw)); v {
-	case "":
-		return "", nil
-	case "SHA", "SHA-256", "SHA-384", "SHA-512":
-		return v, nil
-	case "MD5":
-		return "", errors.New("auth_protocol MD5 is cryptographically broken; use SHA or SHA-256")
-	default:
-		return "", fmt.Errorf("unknown auth_protocol %q; allowed: SHA, SHA-256, SHA-384, SHA-512", raw)
-	}
-}
-
-func normalizePrivProtocol(raw string) (string, error) {
-	switch v := strings.ToUpper(strings.TrimSpace(raw)); v {
-	case "":
-		return "", nil
-	case "AES", "AES-192", "AES-256":
-		return v, nil
-	case "DES":
-		return "", errors.New("priv_protocol DES is cryptographically broken; use AES or AES-256")
-	default:
-		return "", fmt.Errorf("unknown priv_protocol %q; allowed: AES, AES-192, AES-256", raw)
-	}
-}
-
-func (c *Config) validateTargets() error {
-	type hostPort struct {
-		host string
-		port int
-	}
-	seen := make(map[hostPort]int, len(c.Targets))
-	for i, t := range c.Targets {
-		if t.Host == "" {
-			return fmt.Errorf("targets[%d].host must not be empty", i)
-		}
-		if t.Port < 1 || t.Port > 65535 {
-			return fmt.Errorf("targets[%d].port %d is out of range [1, 65535]", i, t.Port)
-		}
-		key := hostPort{host: t.Host, port: t.Port}
-		if first, dup := seen[key]; dup {
-			return fmt.Errorf("targets[%d] duplicates targets[%d] (host:port %s:%d)", i, first, t.Host, t.Port)
-		}
-		seen[key] = i
-		for k := range t.Labels {
-			if k == "" {
-				return fmt.Errorf("targets[%d]: label key must not be empty", i)
-			}
-		}
-	}
-	return nil
-}
-
-// validateFederation enforces LD-15–LD-20 constraints on the federation config.
-func (c *Config) validateFederation() error {
-	if c.Federation.SpokeTimeout < 0 {
-		return fmt.Errorf("federation.spoke_timeout must be >= 0 (0 uses the default of 3× discovery.interval)")
-	}
-	switch c.Federation.Role {
-	case RoleStandalone, RoleUncoordinated:
-		// no extra required fields
-	case RoleSpoke:
-		if c.Federation.Spoke.SpokeID == "" {
-			return errors.New("federation.spoke.spoke_id is required for spoke role")
-		}
-		if err := validateHTTPSEndpoint(c.Federation.Spoke.HubURL); err != nil {
-			return fmt.Errorf("federation.spoke.hub_url: %w", err)
-		}
-		if c.Federation.Spoke.TLSCACert == "" || c.Federation.Spoke.TLSCert == "" || c.Federation.Spoke.TLSKey == "" {
-			return errors.New("federation.spoke.tls_ca_cert, tls_cert, and tls_key are all required for spoke role (LD-20)")
-		}
-		for _, pair := range []struct{ field, path string }{
-			{"federation.spoke.tls_ca_cert", c.Federation.Spoke.TLSCACert},
-			{"federation.spoke.tls_cert", c.Federation.Spoke.TLSCert},
-			{"federation.spoke.tls_key", c.Federation.Spoke.TLSKey},
-		} {
-			if _, err := os.Stat(pair.path); err != nil {
-				return fmt.Errorf("%s: %w", pair.field, err)
-			}
-		}
-	case RoleHub:
-		if c.Federation.Hub.TLSCACert == "" || c.Federation.Hub.TLSCert == "" || c.Federation.Hub.TLSKey == "" {
-			return errors.New("federation.hub.tls_ca_cert, tls_cert, and tls_key are all required for hub role (LD-20)")
-		}
-		for _, pair := range []struct{ field, path string }{
-			{"federation.hub.tls_ca_cert", c.Federation.Hub.TLSCACert},
-			{"federation.hub.tls_cert", c.Federation.Hub.TLSCert},
-			{"federation.hub.tls_key", c.Federation.Hub.TLSKey},
-		} {
-			if _, err := os.Stat(pair.path); err != nil {
-				return fmt.Errorf("%s: %w", pair.field, err)
-			}
-		}
-		if c.Federation.Hub.MaxGraphDevices < 0 {
-			return errors.New("federation.hub.max_graph_devices must be >= 0 (0 = unlimited)")
-		}
-		if c.Federation.Hub.MaxGraphEdges < 0 {
-			return errors.New("federation.hub.max_graph_edges must be >= 0 (0 = unlimited)")
-		}
-		if c.Federation.Hub.MinPushInterval < 0 {
-			return errors.New("federation.hub.min_push_interval must be >= 0 (0 disables the rate limit)")
-		}
-		if c.Federation.Hub.MinPushInterval > 0 && c.Federation.SpokeTimeout > 0 && c.Federation.Hub.MinPushInterval >= c.Federation.SpokeTimeout {
-			return fmt.Errorf("federation.hub.min_push_interval (%s) must be < federation.spoke_timeout (%s) or spokes will be evicted before they can re-push", c.Federation.Hub.MinPushInterval, c.Federation.SpokeTimeout)
-		}
-		// LD-18: spoke_timeout shorter than 2× the discovery interval causes
-		// spokes to be spuriously evicted before they have completed two cycles.
-		if c.Federation.SpokeTimeout < 2*c.Discovery.Interval {
-			return fmt.Errorf("federation.spoke_timeout (%s) must be >= 2× discovery.interval (%s) to prevent spurious eviction (LD-18)", c.Federation.SpokeTimeout, c.Discovery.Interval)
-		}
-	default:
-		return fmt.Errorf("federation.role must be standalone, uncoordinated, spoke, or hub; got %q", c.Federation.Role)
-	}
-	seen := make(map[string]int)
-	for i, link := range c.Federation.KnownInterDomainLinks {
-		if link.LocalDevice == "" || link.LocalPort == "" || link.RemoteDevice == "" || link.RemotePort == "" {
-			return fmt.Errorf("federation.known_inter_domain_links[%d]: local_device, local_port, remote_device, and remote_port are all required", i)
-		}
-		if link.LocalDevice == link.RemoteDevice {
-			return fmt.Errorf("federation.known_inter_domain_links[%d]: local_device and remote_device must differ", i)
-		}
-		fwd := link.LocalDevice + "|" + link.LocalPort + "|" + link.RemoteDevice + "|" + link.RemotePort
-		rev := link.RemoteDevice + "|" + link.RemotePort + "|" + link.LocalDevice + "|" + link.LocalPort
-		for _, key := range []string{fwd, rev} {
-			if seenIdx, ok := seen[key]; ok {
-				return fmt.Errorf("federation.known_inter_domain_links[%d]: duplicate of entry [%d]", i, seenIdx)
-			}
-		}
-		seen[fwd] = i
-		seen[rev] = i
-	}
-	return nil
-}
-
-func (c *Config) validateFDB() error {
-	if c.Modules.FDB.MaxVlans < 1 {
-		return errors.New("fdb.max_vlans must be at least 1")
-	}
-	if c.Modules.FDB.MaxVlans > 4096 {
-		return errors.New("fdb.max_vlans must be at most 4096")
-	}
-	return nil
-}
-
-// scopableModules is the canonical set of per-target-scopable module names,
-// matching the module dispatch table in internal/app/cycle.go and the toggles
-// on ModulesConfig. snmp/arp are intentionally excluded: snmp is the transport
-// (always required to reach a device) and arp is enrichment driven by fdb, not
-// an independently dispatched edge source.
-var scopableModules = map[string]struct{}{
-	"lldp":    {},
-	"cdp":     {},
-	"fdb":     {},
-	"ospf":    {},
-	"bgp":     {},
-	"isis":    {},
-	"mpls_te": {},
-}
-
-// moduleGloballyEnabled reports whether module m is enabled via modules.<m>.enabled.
-// m is assumed to be one of scopableModules (validated upstream).
-func (c *Config) moduleGloballyEnabled(m string) bool {
-	switch m {
-	case "lldp":
-		return c.Modules.LLDP.Enabled
-	case "cdp":
-		return c.Modules.CDP.Enabled
-	case "fdb":
-		return c.Modules.FDB.Enabled
-	case "ospf":
-		return c.Modules.OSPF.Enabled
-	case "bgp":
-		return c.Modules.BGP.Enabled
-	case "isis":
-		return c.Modules.ISIS.Enabled
-	case "mpls_te":
-		return c.Modules.MPLSTE.Enabled
-	default:
-		return false
-	}
-}
-
-// targetOverride is the parsed form of one TargetOverride: the CIDR is parsed
-// into a *net.IPNet once at validate() time, the prefix length is cached for
-// the longest-prefix sort, and the module set is materialised into a lookup map.
-type targetOverride struct {
-	net       *net.IPNet
-	prefixLen int
-	modules   map[string]bool
-}
-
-// targetOverrideResolver answers "which modules may run against this IP" by
-// longest-prefix match over the configured overrides. Built once during
-// validate(); read-only and safe for concurrent use thereafter. It mirrors the
-// most-specific-CIDR-wins model in internal/credentials (Resolver.cidrs).
-type targetOverrideResolver struct {
-	overrides []targetOverride
-}
-
-// buildTargetOverrideResolver parses the overrides and sorts them
-// longest-prefix-first so ModulesForIP returns the most-specific match on its
-// first hit. Tiebreak: when two overrides have an equal prefix length, the
-// first-declared (lower config index) wins — sort.SliceStable preserves config
-// order for equal keys, and validate() rejects exact-duplicate CIDRs so an
-// ambiguous equal-prefix overlap of the *same* network can never silently
-// shadow. CIDRs are assumed already validated by validateTargetOverrides.
-func buildTargetOverrideResolver(overrides []TargetOverride) *targetOverrideResolver {
-	if len(overrides) == 0 {
-		return nil
-	}
-	r := &targetOverrideResolver{overrides: make([]targetOverride, 0, len(overrides))}
-	for _, o := range overrides {
-		_, n, err := net.ParseCIDR(o.CIDR)
-		if err != nil {
-			continue // unreachable: validated upstream.
-		}
-		ones, _ := n.Mask.Size()
-		mods := make(map[string]bool, len(o.Modules))
-		for _, m := range o.Modules {
-			mods[m] = true
-		}
-		r.overrides = append(r.overrides, targetOverride{net: n, prefixLen: ones, modules: mods})
-	}
-	// Most-specific CIDR wins — longest prefix first. SliceStable keeps
-	// config order for equal prefix lengths (first-declared tiebreak).
-	sort.SliceStable(r.overrides, func(i, j int) bool {
-		return r.overrides[i].prefixLen > r.overrides[j].prefixLen
-	})
-	return r
-}
-
-// ModulesForIP returns the set of module names permitted to run against ip,
-// resolved by longest-prefix match over discovery.target_overrides. The
-// returned map is the override's module set (already intersected at config-load
-// time only against the canonical/enabled checks in validation — callers must
-// still AND against the live modules.<m>.enabled gate). When no override
-// matches ip, it returns (nil, false), the signal to run ALL globally-enabled
-// modules (today's unchanged default). The returned map must be treated as
-// read-only.
-func (c *Config) ModulesForIP(ip net.IP) (allowed map[string]bool, matched bool) {
-	r := c.Discovery.overrideResolver
-	if r == nil {
-		return nil, false
-	}
-	for _, o := range r.overrides {
-		if o.net.Contains(ip) {
-			return o.modules, true
-		}
-	}
-	return nil, false
-}
-
-// BuildTargetOverrideResolver validates discovery.target_overrides and rebuilds
-// the cached longest-prefix resolver consulted by ModulesForIP. Load() calls
-// this during validate(); it is also exported so code that constructs a Config
-// programmatically (e.g. tests) can populate the resolver without re-parsing a
-// YAML file. Returns the same errors as the load-time validation path.
-func (c *Config) BuildTargetOverrideResolver() error {
-	return c.validateTargetOverrides()
-}
-
-// validateTargetOverrides enforces the issue #74 invariants: every cidr parses,
-// every referenced module is a recognised name AND enabled globally, and the
-// modules list is non-empty. On success it builds the cached resolver.
-func (c *Config) validateTargetOverrides() error {
-	if len(c.Discovery.TargetOverrides) == 0 {
-		c.Discovery.overrideResolver = nil
-		return nil
-	}
-	seenCIDR := make(map[string]int, len(c.Discovery.TargetOverrides))
-	for i, o := range c.Discovery.TargetOverrides {
-		_, n, err := net.ParseCIDR(o.CIDR)
-		if err != nil {
-			return fmt.Errorf("discovery.target_overrides[%d]: invalid cidr %q: %w", i, o.CIDR, err)
-		}
-		// Reject exact-duplicate CIDRs: two overrides for the same network have
-		// equal prefix length and would make the winner depend on declaration
-		// order alone — an ambiguity better surfaced as a config error.
-		key := n.String()
-		if first, dup := seenCIDR[key]; dup {
-			return fmt.Errorf("discovery.target_overrides[%d]: duplicate cidr %q (already declared at [%d])", i, o.CIDR, first)
-		}
-		seenCIDR[key] = i
-		// An empty modules list is ambiguous: if an operator wants nothing
-		// polled at a CIDR they should exclude it from discovery.scope instead.
-		if len(o.Modules) == 0 {
-			return fmt.Errorf("discovery.target_overrides[%d] (cidr %q): modules must list at least one module (to poll nothing, remove the CIDR from discovery.scope)", i, o.CIDR)
-		}
-		seenMod := make(map[string]struct{}, len(o.Modules))
-		for _, m := range o.Modules {
-			if _, ok := scopableModules[m]; !ok {
-				return fmt.Errorf("discovery.target_overrides[%d] (cidr %q): unknown module %q (allowed: bgp, cdp, fdb, isis, lldp, mpls_te, ospf)", i, o.CIDR, m)
-			}
-			if _, dup := seenMod[m]; dup {
-				return fmt.Errorf("discovery.target_overrides[%d] (cidr %q): duplicate module %q", i, o.CIDR, m)
-			}
-			seenMod[m] = struct{}{}
-			// Acceptance criterion: an override may only narrow, never widen —
-			// every referenced module must be enabled globally.
-			if !c.moduleGloballyEnabled(m) {
-				return fmt.Errorf("discovery.target_overrides[%d] (cidr %q): module %q is not enabled globally (set modules.%s.enabled: true or remove it from the override)", i, o.CIDR, m, m)
-			}
-		}
-	}
-	c.Discovery.overrideResolver = buildTargetOverrideResolver(c.Discovery.TargetOverrides)
-	return nil
-}
-
-func cidrContainsAny(nets []*net.IPNet, ip net.IP) bool {
-	for _, n := range nets {
-		if n.Contains(ip) {
-			return true
-		}
-	}
-	return false
 }
